+++ CHANGES.md
@@ -7,6 +7,7 @@ v0.2.2 (in progress)
 --------------------
 
 - added `Lazy(object).watch` to monitor changes to a property as a sequence (see #49)
+- added `Sequence#chunk`
 - added `Sequence#ofType`
 - deprecated `Lazy.events` in favor of `NodeSequence#on` (note: I should probably rename `NodeSequence` to `DOMSequence`)